Clean up mdx-content functions
